+++ HarpGuru/Menus/AnimatedMenuContainer/components/index.tsx
@@ -2,10 +2,10 @@ import Animated from 'react-native-reanimated'
 import { Dimensions } from 'react-native'
 import React from 'react'
 
-import type { OverlayMenuContainerProps } from '../types'
+import type { MenuContainerProps } from '../types'
 import { styles } from '../../styles'
 
-export const AnimatedMenuContainer = (props: OverlayMenuContainerProps): React.ReactElement => {
+export const AnimatedMenuContainer = (props: MenuContainerProps): React.ReactElement => {
   const { children, onScreen } = props
 
   const { width: windowWidth, height: windowHeight } = Dimensions.get('window')
@@ -13,7 +13,7 @@ export const AnimatedMenuContainer = (props: OverlayMenuContainerProps): React.R
 
   return (
     <Animated.View
-      style={[styles.overlayMenuContainer, {
+      style={[styles.animatedMenuContainer, {
         transform: [
           { translateX: (onScreen ? 0 : guaranteeOffScreenWidth * -1) }
         ],
