+++ HarpGuru/Menus/AnimatedMenuContainer/types/index.ts
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export type OverlayMenuContainerProps = {
+export type MenuContainerProps = {
   readonly onScreen: boolean
   readonly children: React.ReactNode
 }
